+++ Project_Euler_Problem_1/Project_Euler_Problem_1.csproj
@@ -62,8 +62,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CustomMathematics.cs" />
     <Compile Include="Project_Euler_Problem_1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestClass.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
